pyproject.toml
[project]
name = "fapimagaz"
version = "0.1.0"
description = ""
authors = [
    {name = "Kane-Dan",email = "n3t.fox2@gmail.com "}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "celery (>=5.4.0,<6.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "authlib (>=1.5.1,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
